ref: update private layout components
