+++ src/components/Sidenavbar/SidenavbarComponents/NavItem.tsx
@@ -3,6 +3,8 @@ import { Button, Text } from '@chakra-ui/react';
 import { SystemStyleObject } from '@chakra-ui/system';
 import { ReactElement } from 'react';
 import { NavLink } from 'react-router-dom';
+import { Tooltip } from '@/components/ui/tooltip';
+
 
 type NavItemProps = {
     to: string;
@@ -20,31 +22,33 @@ const NavItem = ({ to, text, icon, ...rest }: NavItemProps) => {
     }
 
     return (
-        <Button
-            asChild
-            w={"full"}
-            p={"14px"}
-            h={"auto"}
-            gap={3}
-            fontWeight={"bold"}
-            fontSize={"14px"}
-            justifyContent={"start"}
-            borderRadius={"16px"}
-            _hover={{ bg: "yellow.400" }}
-            variant={"ghost"}
-            css={sidenavabarStyles}
-            {...rest}>
-            <NavLink
-                to={to}
-                className={({ isActive }) => isActive ? "active" : "pending"}
-                onClick={() => setIsOpen(false)}>
-                {icon}
-                <Text
-                    display={{ base: "block", md: isToggle ? "none" : "" }}>
-                    {text}
-                </Text>
-            </NavLink>
-        </Button>
+        <Tooltip content={text} positioning={{ placement: "right" }} disabled={!isToggle}>
+            <Button
+                asChild
+                w={"full"}
+                p={"14px"}
+                h={"auto"}
+                gap={3}
+                fontWeight={"bold"}
+                fontSize={"14px"}
+                justifyContent={"start"}
+                colorPalette={"navItem"}
+                borderRadius={"16px"}
+                variant={"ghost"}
+                css={sidenavabarStyles}
+                {...rest}>
+                <NavLink
+                    to={to}
+                    className={({ isActive }) => isActive ? "active" : "pending"}
+                    onClick={() => setIsOpen(false)}>
+                    {icon}
+                    <Text
+                        display={{ base: "block", md: isToggle ? "none" : "" }}>
+                        {text}
+                    </Text>
+                </NavLink>
+            </Button>
+        </Tooltip>
     )
 }
