+++ src/layout/PrivateLayout.tsx
@@ -1,26 +1,12 @@
-import { ReactNode } from 'react';
 import { Box, useBreakpointValue } from '@chakra-ui/react';
-import { ColorModeButton } from '@/components/ui/color-mode';
 import Sidenavbar from '@/components/Sidenavbar/Sidenavbar';
-import Topbar from '@/components/DashboardComponents/Topbar';
-import { Outlet } from 'react-router-dom';
 import useAutocloseSidenavbar from '@/hooks/useAutocloseSidenavbar';
 import OverlayToLayout from '@/components/ui/OverlayToLayout';
 import { useToolbarTittle } from '@/hooks/useToolbarTitle';
+import MainContainer from '@/components/PrivateLayoutComponents/MainContainer';
+// import { ColorModeButton } from '@/components/ui/color-mode';
 
-type PrivateLayoutProps = {
-  children?: ReactNode;
-  showTopbar?: boolean;
-  showSidebar?: boolean;
-  contentPadding?: string | object;
-};
-
-const PrivateLayout = ({
-  children,
-  showTopbar = true,
-  showSidebar = true,
-  contentPadding = "0"
-}: PrivateLayoutProps) => {
+const PrivateLayout = () => {
 
   const isMobile = useBreakpointValue({ base: true, md: false });
   useAutocloseSidenavbar();
@@ -32,16 +18,7 @@ const PrivateLayout = ({
       position={"relative"}>
       {/* <ColorModeButton pos={"fixed"} bottom={"10px"} left={"10px"} zIndex={1000} border={"md"} borderColor={{ base: "black", _dark: "white" }} /> */}
       <Sidenavbar />
-      <Box
-        as={"main"}
-        position={isMobile ? "absolute" : "relative"} overflowY={"auto"}
-        w={"full"}
-        h={"100vh"}
-        p={contentPadding}
-        bg={{ base: "content.light", _dark: "content.dark" }}>
-        {showTopbar && <Topbar />}
-        {children || <Outlet />}
-      </Box>
+      <MainContainer />
       {isMobile && <OverlayToLayout />}
     </Box>
   );
